change to a decorator, and make the dublin core stuff valid for collections
