+++ lib/oai/provider/metadata_format/hyku_dublin_core.rb
@@ -49,7 +49,11 @@ module OAI
 
         def add_work_url(xml, record)
           work_type = record[:has_model_ssim].first.underscore.pluralize
-          work_path = "https://#{Site.instance.account.cname}/concern/#{work_type}/#{record[:id]}"
+          if work_type === 'collections'
+            work_path = "https://#{Site.instance.account.cname}/#{work_type}/#{record[:id]}"
+          else
+            work_path = "https://#{Site.instance.account.cname}/concern/#{work_type}/#{record[:id]}"
+          end
           xml.tag! 'work_url', work_path
         end
